@@ -1,18 +1,18 @@
 package com.pancakeswap.nft.publish.config;
 
+import lombok.RequiredArgsConstructor;
 import org.springframework.beans.factory.InitializingBean;
-import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.context.annotation.Configuration;
 import org.springframework.context.annotation.Lazy;
 import org.springframework.data.mongodb.core.convert.DefaultMongoTypeMapper;
 import org.springframework.data.mongodb.core.convert.MappingMongoConverter;
 
 @Configuration
+@RequiredArgsConstructor
 public class MongoDBConfig implements InitializingBean {
 
-    @Autowired
     @Lazy
-    private MappingMongoConverter mappingMongoConverter;
+    private final MappingMongoConverter mappingMongoConverter;
 
     @Override
     public void afterPropertiesSet() {
