@@ -5,29 +5,27 @@ import com.pancakeswap.nft.publish.model.entity.CollectionInfo;
 import com.pancakeswap.nft.publish.repository.CollectionInfoRepository;
 import com.pancakeswap.nft.publish.repository.CollectionRepository;
 import com.pancakeswap.nft.publish.service.NFTService;
+import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.bson.types.ObjectId;
 import org.springframework.scheduling.annotation.Scheduled;
 import org.springframework.stereotype.Component;
 
-@Component
+import java.util.concurrent.TimeUnit;
+
 @Slf4j
+@Component
+@RequiredArgsConstructor
 public class ListFailedTokens {
 
     private final NFTService nftService;
     private final CollectionRepository collectionRepository;
     private final CollectionInfoRepository collectionInfoRepository;
 
-    public ListFailedTokens(NFTService nftService, CollectionRepository collectionRepository, CollectionInfoRepository collectionInfoRepository) {
-        this.nftService = nftService;
-        this.collectionRepository = collectionRepository;
-        this.collectionInfoRepository = collectionInfoRepository;
-    }
-
-    @Scheduled(fixedDelay = 1000 * 60 * 60, initialDelay = 0)
+    @Scheduled(fixedDelay = 1, timeUnit = TimeUnit.HOURS, initialDelay = 0)
     public void updateTokens() {
         log.info("updateTokens started");
-        for (Collection collection: collectionRepository.findAll()) {
+        for (Collection collection : collectionRepository.findAll()) {
             CollectionInfo info = collectionInfoRepository.findByCollectionId(new ObjectId(collection.getId()));
             if (info != null && info.getFailedIds() != null) {
                 String[] ids = info.getFailedIds().split(",");
