@@ -1,18 +1,16 @@
 package com.pancakeswap.nft.publish.model.sc;
 
-import lombok.AllArgsConstructor;
 import lombok.Data;
-import lombok.NoArgsConstructor;
+import lombok.RequiredArgsConstructor;
 
 import java.math.BigInteger;
 
 @Data
-@NoArgsConstructor
-@AllArgsConstructor
+@RequiredArgsConstructor
 public class NftInfo {
 
-    private BigInteger prototype;
-    private BigInteger quality;
-    private BigInteger lv;
+    private final BigInteger prototype;
+    private final BigInteger quality;
+    private final BigInteger lv;
 
 }
